Create core dump directory
The core dump directory is set via /proc/sys/kernel/core_pattern and can't be
overridden in the container so it inherits its value from the dea. This
is fine except the dea sets the directory to /var/vcap/sys/cores which
does not exist inside the container. This means that core dumps (which
some buildpacks want to be able to access) cannot be accessed inside the
container. This change adds a matching /var/vcay/sys/cores directory
inside the container so that core dumps work. Note that this still
prevents core dumps by default because the standard warden configuration sets the
max core dump size to 0, but this can (already) be seperately overriden in
the warden's linux.yml.