Fix stacks failures from OS X vs Linux cases
On OS X, the filesystem is case-preserving but not case-senstitive. In
the previous commit we said "require 'Tempfile'", and OS X was able to
find 'tempfile' instead.

Linux, which is not the devil in this respect, is correctly both
case-preserving and case-sensitive. When presented with "Tempfile" it
will fail even if "tempfile" is present.

This patch gives the corrected capitalisation.

[#120114863]

Signed-off-by: Jacques Chester <jchester@pivotal.io>